# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: RTAP API
  version: 1.0.0
  description: Real-Time Analytics Platform — event ingest and daily metrics API.

servers:
  - url: http://localhost:8080
    description: Local development

paths:

  /api/v1/auth/login:
    post:
      summary: Obtain a JWT token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: Esun
              password: Esunadmin
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/events:
    post:
      summary: Ingest an event
      tags: [Events]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
            example:
              eventType: login
              userId: u_123
              sessionId: s_abc
              payload:
                ua: Mozilla/5.0
      responses:
        '200':
          description: Event persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error (eventType is blank)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT
        '429':
          description: Rate limit exceeded (60 req/min per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List events (paginated)
      tags: [Events]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPage'
        '401':
          description: Missing or invalid JWT

  /api/v1/metrics/daily:
    get:
      summary: Daily metric series
      tags: [Metrics]
      parameters:
        - in: query
          name: name
          required: true
          description: Metric name (e.g. DAU, EVENTS_TOTAL)
          schema:
            type: string
            example: DAU
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-07"
      responses:
        '200':
          description: Metric series
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyMetricResponse'
        '400':
          description: "`to` is before `from`"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    TokenResponse:
      type: object
      properties:
        token:
          type: string

    EventRequest:
      type: object
      required: [eventType]
      properties:
        eventType:
          type: string
        userId:
          type: string
        sessionId:
          type: string
        payload:
          type: object
          additionalProperties: true

    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventType:
          type: string
        userId:
          type: string
        sessionId:
          type: string
        payload:
          type: object
        occurredAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time

    EventPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    DailyMetricPoint:
      type: object
      properties:
        metricDate:
          type: string
          format: date
        metricName:
          type: string
        metricValue:
          type: number

    DailyMetricResponse:
      type: object
      properties:
        name:
          type: string
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        points:
          type: array
          items:
            $ref: '#/components/schemas/DailyMetricPoint'

    ErrorResponse:
      type: object
      properties:
        error:
          type: string